ocr/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
pydantic==2.9.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
langgraph==0.0.20
langchain==0.1.0
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
pymupdf==1.23.8
pdfplumber==0.10.3
pdf2image==1.16.3

# OCR Dependencies
typhoon-ocr
azure-ai-formrecognizer==3.3.0
pytesseract==0.3.10
easyocr==1.7.0
opencv-python==4.8.1.78
numpy==1.26.4
requests==2.31.0  